fix(bridge): extract archive and worktree cleanup collaborators
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
